// src/app/api/models/company-profile.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReceiptLanguageEnum } from '../models/receipt-language-enum';
import { ReceiptPaperWidthEnum } from '../models/receipt-paper-width-enum';
export interface CompanyProfile {
  address?: string;
  created_at: string;
  id: number;
  logo?: string | null;
  name: string;
  name_kh?: string;
  phone?: string;
  receipt_footer?: string;
  receipt_header?: string;
  receipt_language?: ReceiptLanguageEnum;
  receipt_paper_width?: ReceiptPaperWidthEnum;
  receipt_show_rate_used?: boolean;
  receipt_show_riel_total?: boolean;
  receipt_show_seller?: boolean;
  updated_at: string;
  vat_tin?: string;
}